refactor(reviews): move TotalScoreCard to its own file
It was colocated with SubmittedReviewView only because both the
submitted and editable branches needed to import from one place, but
the component itself has no submitted-specific logic — it just sums
numeric answers and renders a Surface.
