+++ apps/app/src/components/decisions/Review/SubmittedReviewView.tsx
@@ -4,7 +4,6 @@ import {
   isOverallRecommendationField,
   parseSchemaOptions,
 } from '@op/common/client';
-import { Surface } from '@op/ui/Surface';
 import type { ReactNode } from 'react';
 
 import { useTranslations } from '@/lib/i18n';
@@ -12,7 +11,7 @@ import { useTranslations } from '@/lib/i18n';
 import { FieldHeader } from '../forms/FieldHeader';
 import { compileRubricSchema } from '../forms/rubric';
 import type { FieldDescriptor } from '../forms/types';
-import { getCriteria, inferCriterionType } from '../rubricTemplate';
+import { inferCriterionType } from '../rubricTemplate';
 
 export function SubmittedReviewView({
   rubricTemplate,
@@ -47,41 +46,6 @@ export function SubmittedReviewView({
         </ResultSection>
       )}
     </div>
-  );
-}
-
-export function TotalScoreCard({
-  rubricTemplate,
-  values,
-}: {
-  rubricTemplate: RubricTemplateSchema;
-  values: Record<string, unknown>;
-}) {
-  const t = useTranslations();
-  const criteria = getCriteria(rubricTemplate);
-
-  const totalScore = criteria.reduce<number | null>((total, criterion) => {
-    const value = values[criterion.id];
-
-    if (typeof value !== 'number') {
-      return total;
-    }
-
-    return (total ?? 0) + value;
-  }, null);
-
-  return (
-    <Surface
-      variant="filled"
-      className="flex items-start justify-between rounded-lg border-neutral-gray1 p-4"
-    >
-      <span className="text-base text-neutral-charcoal">
-        {t('Total Score')}
-      </span>
-      <span className="text-base text-neutral-black">
-        {totalScore === null ? '–' : totalScore}
-      </span>
-    </Surface>
   );
 }
